+++ capacitor.config.ts
@@ -1,8 +1,8 @@
 import type { CapacitorConfig } from '@capacitor/cli';
 
 const config: CapacitorConfig = {
-	appId: 'com.opennexus.app',
-	appName: 'open-nexus',
+	appId: 'com.openyaske.app',
+	appName: 'OpenYASKE',
 	webDir: 'build'
 };